git-svn-id: svn://192.168.219.100/torrent-auto-downloader@817 eb2df1e2-efe3-6949-8084-3fc849c3e4eb

+++ trunk/solution_private/TorrentAutoDownloader/src/TorrentAD/src/main/java/kr/co/darkkaiser/torrentad/website/board/AbstractWebSiteBoardItem.java
@@ -9,7 +9,7 @@ import org.jsoup.helper.StringUtil;
 public abstract class AbstractWebSiteBoardItem implements WebSiteBoardItem {
 
 	// 식별자
-	private int identifier;
+	private long identifier;
 
 	// 제목
 	private String title;
@@ -37,7 +37,7 @@ public abstract class AbstractWebSiteBoardItem implements WebSiteBoardItem {
 	}
 
 	@Override
-	public int getIdentifier() {
+	public long getIdentifier() {
 		return this.identifier;
 	}
