+++ trunk/solution_private/TorrentAutoDownloader/src/TorrentAD/src/main/java/kr/co/darkkaiser/torrentad/website/board/WebSiteBoardItem.java
@@ -4,7 +4,7 @@ import java.util.Date;
 
 public interface WebSiteBoardItem {
 
-	int getIdentifier();
+	long getIdentifier();
 
 	String getTitle();
